linux-gen: crc: avoid undefined behavior in bit shift in reflect_u32()
Avoid undefined behavior in bit shift in reflect_u32() by casting to
an unsigned integer of sufficient width before the shift operation.

Fixes GCC undefined sanitizer error:

odp_hash_crc_gen.c:90:16: runtime error: left shift of 237 by 24 places cannot be represented in type 'int'

+++ platform/linux-generic/odp_hash_crc_gen.c
@@ -87,7 +87,7 @@ static inline uint32_t reflect_u32(uint32_t u32)
 	u8[2] = reflect_u8((u32 & 0x0000ff00u) >> 8);
 	u8[3] = reflect_u8(u32 & 0xffu);
 
-	return (u8[3] << 24) | (u8[2] << 16) | (u8[1] << 8) | u8[0];
+	return ((uint32_t)u8[3] << 24) | (u8[2] << 16) | (u8[1] << 8) | u8[0];
 }
 
 /* Reflect 24 bits */
